backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
python-dotenv==1.0.0
openaq==1.0.0
geopy==2.4.1
scipy==1.11.4

